server/experimental/main.py
import os
from dotenv import load_dotenv
load_dotenv()
from langchain.utilities import SQLDatabase
from langchain.llms import OpenAI
from langchain_experimental.sql import SQLDatabaseChain

# print(os.getenv('OPENAI_API_KEY'))

db = SQLDatabase.from_uri("sqlite:///Chinook.db")
llm = OpenAI(temperature=0, verbose=True)
db_chain = SQLDatabaseChain.from_llm(llm, db, verbose=True)

db_chain.run("Who is the most favorited Artist?")
